SpaceObject: add isActive property
The space field only removed space objects out of bounds. It was not
possible to (efficiently) remove specific space objects.

Now, each space object has an "isActive" property. That means a space
object should be on the space field as long as it's active and not out
of bounds. It also allows other objects or methods to remove specific
space objects from the field by simply deactivating it.

e.g. This will be useful later for deactivating missiles or asteroids
after they collide (and explode).