refactor(Explosion): rename based on new abstraction
The basic idea is that an Explosion has 3 states:
    1) Not triggered (it's just bomb)
    2) Triggered and not finished (it's exploding)
    3) Finished (it disappeared)

The names "trigger", "triggerTime", "wasTriggered", and "hasEnded" were
a bit confusing to represent this 3-state model, since all of them
mentions trigger except the latter one.

So now they were renamed to "start", "startTime", "hasStarted", and
"hasEnded".

The wording of the assignment suggested using a property called
"is_triggered", which influentiated the choice of the previous names.
Now the abstraction isn't about "triggering" an explosion, it's about
"starting" it.