fix(Queue): use poll() instead of remove()
Occasionally the program would throw an "NoSuchElementException" when
trying to remove a command from the player commands' Queue.

The verification whether the queue was empty or not was done manually
and, for some reason, it still tried to remove from an empty queue. It
sounds similar to a concurrency problem, but I can't see how it could've
happened.

The solution found was to use the method Queue.poll() instead of
Queue.remove(). It already checks if the queue is empty and returns null
if that's the case.