fix(GameEngine): read all commands from the Queue
TL;DR it guarantees that every Player Command received will be read on
the next tick of the Game Engine.

The game only reads 1 command per update cycle/tick. This means that if
the rate at which commands are received is faster than the update cycle,
the queue would start to fill up. I.e. there would be a delay between
the user commands and the ship's movement.

For example: if the queue size is 5, it would take 5 update cycles for
the game to actually read the latest command received from the user
(since it only reads one command per cycle). As the queue size
increases, it would take increasingly more cycles for the game engine to
catch up with the user's command inputs.

The solution is to, at each update cycle (tick), read ALL the commands
currently available on the queue. That means all the commands received
are guaranteed to be read on the very next update cycle of the game.