fix(Vector2D): avoid division by zero
Given a Vector2D object, the unit vector was implemented as the result
of the division between itself and its own magnitude. That fails when
the vector has 0 magnitude, i.e. when it's the null vector.

The solution was to define the unit vector of a null vector as a null
vector.