refactor: rename ScoreHistory port and adapters
Renamed ScoreHistory to ScoreArchiver, LeaderboardHistory to
LeaderboardArchiver, and ScoreHistory to ScoreArchiver.

The save() method used to call load() to load the scoreboard, add the
new score to the scoreboard loaded, and then save the new scoreboard
generated. Now the proccess was separated into different methods:

The load() method is supposed to be called externally to load the
Scoreboard.json and parse it as a list of scores, which is then saved
internally. If load() was not called, this internal list of scores is
defaulted as an empty list.

The submit() method receives a score and adds it to the internally saved
list of scores (not in place), generating a new list of scores. Then,
the save() method is called with this new list of scores as parameter.

The save() method receives a list of scores and generates a JSON string,
which is then saved to the Scoreboard.json file.